# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DSS Backend API",
        "description": "KUSRC - DSS Backend API V0.0.3 (Update 18-09-2024 04:20AM)",
        "version": "0.0.3"
    },
    "servers": [
        {
            "url": "http://172.16.130.54:8000"
        }
    ],
    "paths": {
...
        "/api/getMoveHistories": {
            "get": {
                "tags": [
                    "Move Histories"
                ],
                "summary": "GET: Move Inventories Histories Data with Pagination",
                "operationId": "0c02ba4c975098455ff562a61e8ae8ca",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/getMoveHistories/{id}": {
            "get": {
                "tags": [
                    "Move Histories"
                ],
                "summary": "GET: Move Inventories Histories Data by ID",
                "operationId": "e90c2e9fe5ed8d39e2d8f2221f2111ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Move Inventory History ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/getAllMoveHistories": {
            "get": {
                "tags": [
                    "Move Histories"
                ],
                "summary": "GET: All Move Inventories Histories Data",
                "operationId": "b9797fb4d354a2135314acae6b39c7cd",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/getAllMoveHistories/{id}": {
            "get": {
                "tags": [
                    "Move Histories"
                ],
                "summary": "GET: All Move Inventories Histories Data by ID",
                "operationId": "66ff820ffcde61c57ebc3f74b9733182",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Move Inventory History ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/postMoveHistories": {
            "post": {
                "tags": [
                    "Move Histories"
                ],
                "summary": "POST: Create Move Inventory History Data",
                "operationId": "4671dde8f211ca1941565d1f0a388add",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "inv_id": {
                                        "type": "number",
                                        "example": 1
                                    },
                                    "detail": {
                                        "type": "string",
                                        "example": "ทดสอบ"
                                    },
                                    "building_id": {
                                        "type": "number",
                                        "example": 2
                                    },
                                    "floor": {
                                        "type": "number",
                                        "example": 4
                                    },
                                    "room_number": {
                                        "type": "string",
                                        "example": "ห้องอบรมคอมพิวเตอร์"
                                    },
                                    "sub_division_id": {
                                        "type": "number",
                                        "example": 1801
                                    },
                                    "owner_emp_id": {
                                        "type": "number",
                                        "example": 15404
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
...
}